asko_import: Split out object_store and fields modules
Move ObjectStore to object_store module and the field classes to fields
module.
